Add schema and treat numeric data as numeric.
If we want to add PCA we need to be aware of some limitations. PCA can only
be performed on numeric data. However, d3.csv loads all variables as
strings. To work around this we now require one extra header line in the
csv file, which has either "factor" or "numeric" for each of the variables.
We use this line to transform the numeric variables to actual numbers.